universe & pointers
